+++ cloud-minecraft/cloud-fabric/src/main/java/cloud/commandframework/fabric/FabricExecutor.java
@@ -32,6 +32,7 @@ import cloud.commandframework.exceptions.NoSuchCommandException;
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import com.mojang.logging.LogUtils;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.concurrent.CompletionException;
@@ -44,16 +45,13 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.ComponentUtils;
 import net.minecraft.network.chat.HoverEvent;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.network.chat.TextComponent;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.checkerframework.checker.nullness.qual.NonNull;
+import org.slf4j.Logger;
 
 final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Command<S> {
+    private static final Logger LOGGER = LogUtils.getLogger();
 
-    private static final Logger LOGGER = LogManager.getLogger();
-
-    private static final Component NEWLINE = new TextComponent("\n");
+    private static final Component NEWLINE = Component.literal("\n");
     private static final String MESSAGE_INTERNAL_ERROR = "An internal error occurred while attempting to perform this command.";
     private static final String MESSAGE_NO_PERMS =
             "I'm sorry, but you do not have permission to perform this command. "
@@ -79,6 +77,7 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
         final S source = ctx.getSource();
         final String input = ctx.getInput().substring(ctx.getLastChild().getNodes().get(0).getRange().getStart());
         final C sender = this.manager.commandSourceMapper().apply(source);
+
         this.manager.executeCommand(sender, input).whenComplete((result, throwable) -> {
             if (throwable == null) {
                 return;
@@ -99,8 +98,8 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
                     (InvalidSyntaxException) throwable,
                     (c, e) -> this.sendError.accept(
                             source,
-                            new TextComponent("Invalid Command Syntax. Correct command syntax is: ")
-                                    .append(new TextComponent(String.format("/%s", e.getCorrectSyntax()))
+                            Component.literal("Invalid Command Syntax. Correct command syntax is: ")
+                                    .append(Component.literal(String.format("/%s", e.getCorrectSyntax()))
                                             .withStyle(style -> style.withColor(ChatFormatting.GRAY)))
                     )
             );
@@ -109,21 +108,21 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
                     sender,
                     InvalidCommandSenderException.class,
                     (InvalidCommandSenderException) throwable,
-                    (c, e) -> this.sendError.accept(source, new TextComponent(throwable.getMessage()))
+                    (c, e) -> this.sendError.accept(source, Component.literal(throwable.getMessage()))
             );
         } else if (throwable instanceof NoPermissionException) {
             this.manager.handleException(
                     sender,
                     NoPermissionException.class,
                     (NoPermissionException) throwable,
-                    (c, e) -> this.sendError.accept(source, new TextComponent(MESSAGE_NO_PERMS))
+                    (c, e) -> this.sendError.accept(source, Component.literal(MESSAGE_NO_PERMS))
             );
         } else if (throwable instanceof NoSuchCommandException) {
             this.manager.handleException(
                     sender,
                     NoSuchCommandException.class,
                     (NoSuchCommandException) throwable,
-                    (c, e) -> this.sendError.accept(source, new TextComponent(MESSAGE_UNKNOWN_COMMAND))
+                    (c, e) -> this.sendError.accept(source, Component.literal(MESSAGE_UNKNOWN_COMMAND))
             );
         } else if (throwable instanceof ArgumentParseException) {
             this.manager.handleException(
@@ -132,14 +131,14 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
                     (ArgumentParseException) throwable,
                     (c, e) -> {
                         if (throwable.getCause() instanceof CommandSyntaxException) {
-                            this.sendError.accept(source, new TextComponent("Invalid Command Argument: ")
-                                    .append(new TextComponent("")
+                            this.sendError.accept(source, Component.literal("Invalid Command Argument: ")
+                                    .append(Component.literal("")
                                             .append(ComponentUtils
                                                     .fromMessage(((CommandSyntaxException) throwable.getCause()).getRawMessage()))
                                             .withStyle(ChatFormatting.GRAY)));
                         } else {
-                            this.sendError.accept(source, new TextComponent("Invalid Command Argument: ")
-                                    .append(new TextComponent(throwable.getCause().getMessage())
+                            this.sendError.accept(source, Component.literal("Invalid Command Argument: ")
+                                    .append(Component.literal(throwable.getCause().getMessage())
                                             .withStyle(ChatFormatting.GRAY)));
                         }
                     }
@@ -151,7 +150,7 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
                     (CommandExecutionException) throwable,
                     (c, e) -> {
                         this.sendError.accept(source, this.decorateHoverStacktrace(
-                                new TextComponent(MESSAGE_INTERNAL_ERROR),
+                                Component.literal(MESSAGE_INTERNAL_ERROR),
                                 throwable.getCause(),
                                 sender
                         ));
@@ -164,7 +163,7 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
             );
         } else {
             this.sendError.accept(source, this.decorateHoverStacktrace(
-                    new TextComponent(MESSAGE_INTERNAL_ERROR),
+                    Component.literal(MESSAGE_INTERNAL_ERROR),
                     throwable,
                     sender
             ));
@@ -183,9 +182,9 @@ final class FabricExecutor<C, S extends SharedSuggestionProvider> implements Com
         return input.withStyle(style -> style
                 .withHoverEvent(new HoverEvent(
                         HoverEvent.Action.SHOW_TEXT,
-                        new TextComponent(stackTrace)
+                        Component.literal(stackTrace)
                                 .append(NEWLINE)
-                                .append(new TextComponent("    Click to copy")
+                                .append(Component.literal("    Click to copy")
                                         .withStyle(s2 -> s2.withColor(ChatFormatting.GRAY).withItalic(true)))
                 ))
                 .withClickEvent(new ClickEvent(